--- Capstone_API/Capstone_API/Repository/Interface/ILoginRepository.cs
@@ -1,12 +1,0 @@
-﻿using Capstone_API.Models;
-using System.Collections.Generic;
-using System.Xml.Serialization;
-
-namespace Capstone_API.Repository.Interface
-{
-    public interface ILoginRepository
-    {
-
-        public Account GetAccount(string phone, string password);
-    }
-}
